+++ pyproject.toml
@@ -11,9 +11,10 @@ aiosmtplib = "^3.0.2"
 alembic = "^1.13.2"
 asyncpg = "^0.29.0"
 boto3 = "^1.34.162"
-email-validator = "^2.2.0"
+email-validator = "2.1.2"
 fastapi = "^0.112.1"
 fastapi-cache = "^0.1.0"
+fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
 google-auth-oauthlib = "^1.2.1"
 hiredis = "^3.0.0"
 jmespath = "^1.0.1"
@@ -21,7 +22,7 @@ mutagen = "^1.47.0"
 passlib = "^1.7.4"
 pydantic = "^2.8.2"
 pydantic-settings = "^2.4.0"
-pyjwt = "^2.9.0"
+pyjwt = "2.8.0"
 python = "^3.11"
 python-dotenv = "^1.0.1"
 redis = "^5.0.8"
@@ -66,9 +67,6 @@ max-line-length = 200
 ignore = [
     "E731",
     "F821",
-    "F401",
-    "E501",
-    "F821"
 ]
 exclude =[
     ".git/",
