+++ src/models/albums.py
@@ -1,7 +1,8 @@
 from sqlalchemy import Column, Table, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from src.core.database import Base
+from src.models.base import Base
+
 
 artist_profile_album_association = Table(
     "artist_profile_album_association",
