+++ src/models/beatpacks.py
@@ -1,8 +1,8 @@
 from sqlalchemy import Integer, String, Table, ForeignKey, Column
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from src.core.database import Base
 from src.models.auth import User
+from src.models.base import Base
 from src.models.beats import Beat
