+++ src/models/messages.py
@@ -1,4 +1,4 @@
-from src.core.database import Base
+from src.models.base import Base
 
 
 class Chat(Base):
