+++ src/models/notifications.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src.core.database import Base
+from src.models.base import Base
 
 
 class Notification(Base):
