+++ tests/conftest.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import create_async_engine
 
 from src.api.app import app
 from src.core.config import settings
-from src.core.database import Base
+from src.models.base import Base
 from src.schemas.auth import SRegisterUserRequest, Role, SLoginRequest, SLoginResponse
 
 engine_test = create_async_engine(settings.db.url, echo=True)
